use new component for multiselect search link
