+++ src/components/ClickToSearchLink/ClickToSearchLink.vue
@@ -1,11 +1,8 @@
 <template>
   <span v-if="props.widget.clickToSearch">
-    <button
-      class="text-blue-600 hover:text-blue-700 hover:underline"
-      @click="handleClick"
-    >
+    <a :href="linkUrl">
       <slot />
-    </button>
+    </a>
   </span>
   <span v-else>
     <slot />
@@ -13,54 +10,18 @@
 </template>
 
 <script setup lang="ts">
-import { computed, nextTick } from "vue";
+import { computed } from "vue";
 import { WidgetProps } from "@/types";
-import { useSearchStore } from "@/stores/searchStore";
-import { useRouter } from "vue-router";
+import { toClickToSearchUrl } from "@/helpers/displayUtils";
 
-const props = defineProps<{
+interface Props {
   linkText: string;
   widget: WidgetProps;
-}>();
-
-const searchStore = useSearchStore();
-
-const fieldId = computed((): string => props.widget.fieldTitle);
-const fieldValue = computed((): string => {
-  const cleanedLinkText = props.linkText
-    .trim()
-    .replace("?", "")
-    .replace("...", "");
-
-  return cleanedLinkText.split(" : ").join(",");
-});
-
-const router = useRouter();
-
-async function handleClick() {
-  // this async function will take awhile to run
-  // to give the appearance of a loading state
-  // we'll queue up the search on next tick
-  // and then navigate to a blank page while waiting for the search to complete
-  // once the search is complete, we'll navigate to the search page
-  nextTick(async () => {
-    // clear search
-    searchStore.reset();
-
-    // create a new searchable field filter
-    searchStore.addSearchableFieldFilter(fieldId.value, {
-      value: fieldValue.value,
-    });
-
-    // get the search id
-    const searchId = await searchStore.getSearchId();
-    searchStore.search(searchId);
-
-    // navigate to the search page
-    router.replace(`/search/s/${searchId}`);
-  });
-
-  // go to a blank page while waiting for the search to complete
-  router.push(`/blank`);
 }
+
+const props = defineProps<Props>();
+
+const linkUrl = computed((): string =>
+  toClickToSearchUrl(props.linkText, props.widget)
+);
 </script>
